+++ lib/vagrant/util/subprocess.rb
@@ -73,17 +73,17 @@ module Vagrant
         process.io.stderr = stderr_writer
         process.duplex = true
 
-        # If we're in the installer and the command is NOT included
-        # in the installer (external), then remove the DYLD_IMPORT_PATH
-        # environmental variable to run into linker issues. [GH-2219]
-        if Vagrant.in_installer? && ENV["VAGRANT_ORIGINAL_DYLD_LIBRARY_PATH"]
+        # If we're in an installer on Mac and we're executing a command
+        # in the installer context, then force DYLD_LIBRARY_PATH to look
+        # at our libs first.
+        if Vagrant.in_installer? && Platform.darwin?
           installer_dir = ENV["VAGRANT_INSTALLER_EMBEDDED_DIR"].to_s.downcase
-          if !@command[0].downcase.include?(installer_dir)
-            @logger.info("Command not in the installer. Removing DYLD_LIBRARY_PATH")
+          if @command[0].downcase.include?(installer_dir)
+            @logger.info("Command in the installer. Specifying DYLD_LIBRARY_PATH...")
             process.environment["DYLD_LIBRARY_PATH"] =
-              ENV["VAGRANT_ORIGINAL_DYLD_LIBRARY_PATH"]
+              "#{installer_dir}/lib:#{ENV["DYLD_LIBRARY_PATH"]}"
           else
-            @logger.debug("Command in installer, not touching env vars.")
+            @logger.debug("Command not in installer, not touching env vars.")
           end
         end