+++ CHANGELOG.md
@@ -8,6 +8,8 @@ FEATURES:
 BUG FIXES:
 
   - core: Don't output `\e[0K` anymore on Windows. [GH-2246]
+  - core: Only modify `DYLD_LIBRARY_PATH` on Mac when executing commands
+    in the installer context. [GH-2231]
   - hosts/arch: Vagrant won't crash on Arch anymore. [GH-2233]
 
 ## 1.3.3 (September 18, 2013)